+++ VMWithEditor/UnitTests/State_Machine_Unit_Test/State_Machine_Unit_Test/internal_sytax_state_tests.h
@@ -2,8 +2,9 @@
 #define INTERNAL_SYNTAX_STATE_TEST_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
-bool internal_tests_on_all_state_transitions(unsigned test_step);
-bool unit_test_lexical_analyzer(unsigned test_step);
+bool internal_tests_on_all_state_transitions(size_t test_step);
+bool unit_test_lexical_analyzer(size_t test_step);
 
 #endif	// INTERNAL_SYNTAX_STATE_TEST_H
